refactor: Add parser to utils package
